Allow pinning of the git resource ref
This allows users to specify the version of the git repository that
should be checked out rather than following master. It is designed to be
used in situations where a third-party git repository is part of a
pipeline and you wish to have control over the version available.

It would possibly make sense to extend this to allow users to specify a
file containing a ref to check out as well as allowing the ref to be
specified in the pipeline.